cleaned up raw queries as python functions
