+++ queryPages/custom_query.py
@@ -1,29 +1,20 @@
 import streamlit as st
 import pandas as pd
+from queries.general_query import *
 
 '''
 This page is for running custom queries on the database. need to import the actual queries from queries folder (or extract the placeholders I made here)
 '''
-
 def show(conn):
     st.header("Custom SQL Query")
     
     # Show table schema subsection -----
     if st.checkbox("Show Database Schema"):
-        tables = conn.execute("""
-            SELECT table_name 
-            FROM information_schema.tables 
-            WHERE table_schema = 'main'
-        """).fetchdf()
-        
+        tables = get_all_table_info(conn)
         for _, row in tables.iterrows():
             table = row['table_name']
             st.subheader(f"Table: {table}")
-            columns = conn.execute(f"""
-                SELECT column_name, data_type
-                FROM information_schema.columns
-                WHERE table_name = '{table}' AND table_schema = 'main'
-            """).fetchdf()
+            columns = get_all_table_columns(conn, table)
             st.dataframe(columns)
     
     # Custom query input subsection -----
@@ -46,4 +37,4 @@ def show(conn):
         else:
             st.warning("Please enter a query") 
     
-    # I think that the personal queries are good enough as is -> only other feature I can think of is some way to visualize the graph with arrows and stuff (maybe?)
+    # I think that the personal queries are good enough as is -> only other feature I can think of is some way to visuconn.execalize the graph with arrows and stuff (maybe?)
